refactor(migrations): squash forms migrations into one matching the reworked model
