--- backend/alembic/versions/0eba309024e1_form_reponse_and_answer_models.py
@@ -1,55 +1,0 @@
-"""form reponse and answer models
-
-Revision ID: 0eba309024e1
-Revises: 2b134a51cf44
-Create Date: 2026-08-12 22:40:57.539707
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '0eba309024e1'
-down_revision: Union[str, None] = '2b134a51cf44'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('form_responses',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('form_id', sa.Integer(), nullable=False),
-    sa.Column('user_id', sa.Integer(), nullable=False),
-    sa.Column('submitted_at', sa.DateTime(timezone=True), nullable=True),
-    sa.Column('updated_at', sa.DateTime(timezone=True), nullable=True),
-    sa.ForeignKeyConstraint(['form_id'], ['forms.id'], ondelete='CASCADE'),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('form_id', 'user_id', name='uq_form_response_per_user')
-    )
-    op.create_index(op.f('ix_form_responses_id'), 'form_responses', ['id'], unique=False)
-    op.create_table('form_answers',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('response_id', sa.Integer(), nullable=False),
-    sa.Column('field_id', sa.Integer(), nullable=False),
-    sa.Column('value', sa.JSON(), nullable=False),
-    sa.ForeignKeyConstraint(['field_id'], ['form_fields.id'], ),
-    sa.ForeignKeyConstraint(['response_id'], ['form_responses.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('response_id', 'field_id', name='uq_answer_per_field')
-    )
-    op.create_index(op.f('ix_form_answers_id'), 'form_answers', ['id'], unique=False)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_form_answers_id'), table_name='form_answers')
-    op.drop_table('form_answers')
-    op.drop_index(op.f('ix_form_responses_id'), table_name='form_responses')
-    op.drop_table('form_responses')
-    # ### end Alembic commands ###
