--- backend/alembic/versions/2b134a51cf44_form_and_form_fields_models.py
@@ -1,124 +1,0 @@
-"""form and form fields models
-
-Revision ID: 2b134a51cf44
-Revises: d3e4f5a6b7c8
-Create Date: 2026-08-12 00:46:13.556533
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import postgresql
-
-# revision identifiers, used by Alembic.
-revision: str = '2b134a51cf44'
-down_revision: Union[str, None] = 'd3e4f5a6b7c8'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('forms',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('owner_type', sa.String(length=16), nullable=False),
-    sa.Column('tournament_id', sa.Integer(), nullable=True),
-    sa.Column('chapter_id', sa.Integer(), nullable=True),
-    sa.Column('name', sa.String(length=255), nullable=False),
-    sa.Column('created_by', sa.Integer(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), nullable=True),
-    sa.Column('updated_at', sa.DateTime(timezone=True), nullable=True),
-    sa.CheckConstraint("(owner_type = 'tournament' AND tournament_id IS NOT NULL AND chapter_id IS NULL) OR (owner_type = 'chapter' AND chapter_id IS NOT NULL AND tournament_id IS NULL)", name='ck_form_owner_exclusive'),
-    sa.ForeignKeyConstraint(['chapter_id'], ['alumni_chapters.id'], ondelete='CASCADE'),
-    sa.ForeignKeyConstraint(['created_by'], ['users.id'], ),
-    sa.ForeignKeyConstraint(['tournament_id'], ['tournaments.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_forms_id'), 'forms', ['id'], unique=False)
-    op.create_table('form_fields',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('form_id', sa.Integer(), nullable=False),
-    sa.Column('order', sa.Integer(), nullable=False),
-    sa.Column('label', sa.String(length=255), nullable=False),
-    sa.Column('description', sa.Text(), nullable=True),
-    sa.Column('question_type', sa.String(length=32), nullable=False),
-    sa.Column('field_key', sa.String(length=64), nullable=False),
-    sa.Column('config', sa.JSON(), nullable=True),
-    sa.Column('required', sa.Boolean(), nullable=False),
-    sa.Column('is_archived', sa.Boolean(), nullable=False),
-    sa.ForeignKeyConstraint(['form_id'], ['forms.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('form_id', 'field_key', name='uq_form_field_key')
-    )
-    op.create_index(op.f('ix_form_fields_id'), 'form_fields', ['id'], unique=False)
-    op.alter_column('sheet_configs', 'created_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               nullable=True)
-    op.alter_column('sheet_configs', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               nullable=True)
-    op.alter_column('tournament_events', 'created_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               nullable=True)
-    op.alter_column('tournament_events', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               nullable=True)
-    op.alter_column('tournament_memberships', 'created_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               nullable=True)
-    op.alter_column('tournament_memberships', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               nullable=True)
-    op.alter_column('tournaments', 'created_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               nullable=True)
-    op.alter_column('tournaments', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               nullable=True)
-    op.alter_column('users', 'created_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               nullable=True)
-    op.alter_column('users', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               nullable=True)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('users', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               nullable=False)
-    op.alter_column('users', 'created_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               nullable=False)
-    op.alter_column('tournaments', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               nullable=False)
-    op.alter_column('tournaments', 'created_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               nullable=False)
-    op.alter_column('tournament_memberships', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               nullable=False)
-    op.alter_column('tournament_memberships', 'created_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               nullable=False)
-    op.alter_column('tournament_events', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               nullable=False)
-    op.alter_column('tournament_events', 'created_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               nullable=False)
-    op.alter_column('sheet_configs', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               nullable=False)
-    op.alter_column('sheet_configs', 'created_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               nullable=False)
-    op.drop_index(op.f('ix_form_fields_id'), table_name='form_fields')
-    op.drop_table('form_fields')
-    op.drop_index(op.f('ix_forms_id'), table_name='forms')
-    op.drop_table('forms')
-    # ### end Alembic commands ###
